+++ app/config.py
@@ -6,6 +6,7 @@ from .daos.solution_dao import SolutionDAO
 from .env_vars import ENV_VARS
 from .managers.auth_manager import AuthManager, JWTBearer
 from .managers.firebase_manager import FirebaseManager
+from .managers.ai_analysis_manager import AiAnalysisManager
 from .managers.metadata_manager import MetadataManager
 from .managers.question_manager import QuestionManager
 from .managers.solution_manager import SolutionManager
@@ -19,6 +20,7 @@ class Configs:
     solution_manager: SolutionManager = None
     auth_manager: AuthManager = None
     metadata_manager: MetadataManager = None
+    ai_analysis_manager = None
     question_collection: str = ""
     solution_collection: str = ""
 
@@ -49,8 +51,7 @@ class Configs:
             )
 
             cls.metadata_manager = MetadataManager()
-            cls.question_manager = QuestionManager(cls.question_dao, cls.metadata_manager)
-            cls.solution_manager = SolutionManager(cls.solution_dao)
+
             cls.auth_manager = AuthManager(
                 ENV_VARS.get("APP_SECRET"),
                 ENV_VARS.get("JWT_SIGNITURE"),
@@ -59,6 +60,33 @@ class Configs:
                 ENV_VARS.get("JWT_ISS"),
                 ENV_VARS.get("JWT_AUD")
             )
+
+            ai_ctx = """
+            You are an expert code reviewer. You will be given a coding problem,
+            a solution and the language its written in.
+            Note that the solution may be invalid, empty or incorrect.
+            Analyze the solution briefly but insightfully in 1-2 paragraphs.
+
+            Your feedback should assess correctness, efficiency, and clarity.
+            Mention strengths and suggest improvements if applicable.
+            Point out any errors in the code if applicable.
+            Highlight the worst-case time and space complexity.
+
+            Return the data in a serilized JSON string for python to deserialize with the following fields:
+            - time_complexity: str, the WC time complexity of the solution (e.g. n^2, not O(n^2))
+            - space_complexity: str, the WC space complexity of the solution (e.g. n, not O(n))
+            - feedback: str, the feedback to the solution
+
+            If any information is missing or invalid, ignore it. Do NOT provide a solution nor a possible solution.
+            If rate limit reached return empty string for each field.
+            """
+
+            cls.ai_analysis_manager = AiAnalysisManager(
+                ENV_VARS.get("GEMINI_API_KEY"), ENV_VARS.get("GEMINI_MODEL"), ai_ctx
+            )
+
+            cls.question_manager = QuestionManager(cls.question_dao, cls.metadata_manager)
+            cls.solution_manager = SolutionManager(cls.solution_dao, cls.ai_analysis_manager)
 
         return cls.instance
 
@@ -81,6 +109,10 @@ def get_metadata_service(configs: Annotated[Configs, Depends(init_config)]):
 
 def get_auth_service(configs: Annotated[Configs, Depends(init_config)]):
     return configs.auth_manager
+
+
+def get_ai_analysis_service(configs: Annotated[Configs, Depends(init_config)]):
+    return configs.ai_analysis_manager
 
 
 auth_user_jwt = JWTBearer(Configs().auth_manager)