+++ js/components/question/solution_form.jsx
@@ -46,7 +46,9 @@ export default function SolutionForm(props) {
       time_complexity: timeComplexity,
       space_complexity: spaceComplexity,
       code: code,
-      accepted: accepted
+      accepted: accepted,
+      question_title: props.questionTitle,
+      question_prompt: props.questionPrompt
     };
 
     if (isFormValid(data)) {